pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "pinecone>=8.0.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
]
